+++ docs/changelog.rst
@@ -7,6 +7,9 @@ Next release
 v0.8 (September 9, 2022)
 ------------------------
 
+* Data structure: ``spfd`` (contributed by Andrea Costamagna)
+  `#140 <https://github.com/msoeken/kitty/pull/140>`_
+
 * Data structure: ``ternary_truth_table`` and ``quaternary_truth_table`` (contributed by Siang-Yun Lee and Gianluca Radi)
   `#133 <https://github.com/msoeken/kitty/pull/133>`_